hw: acpi: Factorize ACPI definitions for PCI hotplug
This commit factorizes the code responsible of adding the PCI hotplug
ability to the PCI device in the DSDT table. This way, the code can
now be shared between virt and pc machine types.
